waitingBotton 컴포넌트 분리 이제 함수와 앤드포인트를 만들자
1. 대기중 상태의 친구의 상태를 바꿔주는 버튼(승낙)을 만들어야한다.
 이 작업 중이다
   일단 friendBoard를 만들었다. 만들어야할 상태는
a. 온라인
b. 모두
c. 대기중
현재 여기 작업중
대기중 보드를 따로 만들 필요 없이 friendListState에 따라 동적으로 변하게 만들 것이다.
일단 이미지까지 가져옴
이름 해결 css까지
상태 해결 css까지
버튼 해결 css까지
이후 버튼 컴포넌트를 따로 분류하고 온고잉과 인커밍으로 들어오는 타입에 따라
바뀌게 바꾸었고 CancelButton를 만들었다.

이후 각 버튼의 함수와 앤드포인트를 만들어야함
d 차단
